back-end/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
pymysql>=1.0
passlib[bcrypt]>=1.7
python-dotenv>=0.21
aiomysql
aiofiles
aiohttp
aiosmtplib
email-validator
PyJWT
